[DVO-3104] Allow specifying static whitelists and blacklists at provision time.
When the stack is deployed, the whitelist and blacklist are both initialized
to empty. If they are modified manually, and the stack is updated, any
manual changes are removed, resulting in empty lists.

This change allows CloudFormation to optionally manage the contents of these
lists, so that they are not wiped out on update.